+++ .gitignore
@@ -7,3 +7,4 @@
 ChangeLog
 PyKMIP.egg-info/
 dist/
+htmlcov/
