+++ kmip/tests/unit/core/attributes/test_application_specific_information.py
