+++ kmip/tests/unit/core/misc/test_misc.py
