+++ kmip/tests/unit/core/objects/test_extension_information.py
