+++ tox.ini
@@ -6,11 +6,18 @@ deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
 
 commands =
-    coverage run --source=kmip/ -m pytest --strict {posargs}
-    coverage report
+    coverage run --source=kmip/ --omit=kmip/demos/*,kmip/tests/* -m pytest --strict kmip/tests/unit
+    coverage report -m
+    coverage html
 
 [testenv:pep8]
-commands = flake8 {posargs}
+commands = flake8 kmip/
+
+[testenv:integration]
+# Note: This requires local or remote access to a KMIP appliance or service
+deps = {[testenv]deps}
+commands =
+    py.test --strict kmip/tests/integration
 
 [flake8]
 exclude = .git,.tox,dist,rpmbuild,*.egg-info
